requirements.txt
fyers-apiv3
typer
rich
python-dotenv
